requirements.txt
# Telegram Bot
python-telegram-bot>=13.7

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
crawl4ai>=0.1.0
playwright>=1.41.0

# Data Processing
pandas>=1.3.0
pydantic>=1.8.2

# Storage
python-dotenv>=0.19.0

# Testing
pytest>=6.2.5
pytest-mock>=3.6.1

# Logging
loguru>=0.5.3

# Utilities
tqdm>=4.62.0

# Natural Language Processing
openai>=1.0.0
tiktoken>=0.5.0 